requirements.txt
openai
tqdm
python-dotenv
fire
pandas
requests
datasets
